Property Provider configuration extension method work
Unfortunately the extension methods over the configuration interface would now be more verbose than instantiating the types directly due to all the generic parameters and [lack of partial eval of generic type parameters in C#](http://stackoverflow.com/questions/2345140/extra-generic-parameter-in-generic-extension-methods), so I have removed those.

I have instead added extension methods for the `List<IMessagePropertyProvider<T>>` which enable simpler configuration for the `CompositeMessagePropertyProvider` class which was the majority of the functionality we needed anyway.

The only thing lost here is if you only want a single `FuncMessagePropertyProvider<> ` you'll have to instantiate that class yourself instead of having a nice configuration method overload.